pyproject.toml
[tool.poetry]
name = "movie-reservation"
version = "0.1.0"
description = ""
authors = ["Kirill Dvoretsky <shortykofficial@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
phonenumbers = "^8.13.46"
alembic = "^1.13.3"
psycopg = {extras = ["binary"], version = "^3.2.2"}
pydantic-extra-types = "^2.9.0"
pydantic-settings = "^2.5.2"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
email-validator = "^2.2.0"
black = "^24.8.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
